requirements.txt
streamlit==1.32.2
python-dotenv==1.0.1
PyPDF2==3.0.1
faiss-cpu==1.7.4
sentence-transformers==2.6.1
numpy==1.26.4
requests==2.31.0
